RMU: Erase fields from RedexContext, not just from their class defs
Summary:
RedexContext has logic to prevent the creation of two fields with the same name
and type: https://fburl.com/hh1ppemp. This is a useful check. However,
ObfuscatePass was tripping up over it because RMU was not erasing deleted
fields from RedexContext. ObfuscatePass would examine the DexClass definitions,
observe that some field F was absent, and attempt to assign its name to
another, thus hitting the assert.

This was uncovered during testing of the RemoveUnreadFields pass.

Reviewed By: minjang

Differential Revision: D7151271

fbshipit-source-id: 3a9546dbed59a85d479ec3d29caf9f7a6f9cb684