Fix v124 oatmeal generation segfault
Summary:
the OatDexFileRecord appears to be v131 only but we are still attempting to write it
at 8.0 v124
on an mac this results in a segfault when we attempt to write:
* thread #1, queue = 'com.apple.main-thread', stop reason = EXC_BAD_ACCESS (code=1, address=0x7ffeefc00000)
...
frame #3: 0x0000000100000851 oatmeal`FileHandle::fwrite_impl(this=0x00007ffeefbfcfb8, p=0x00007ffeefbfd071, size=1, count=2811670808) at OatmealUtil.cpp:16
...

on the device we get a message in stderr with
native/redex/tools/oatmeal/OatmealUtil.cpp:60 CHECK(fh.fwrite(buf.ptr, sizeof(char), buf.len) == buf.len) failed.
but the binary doesnt seem to crash even though there is an assert there.

Reviewed By: maxmg22

Differential Revision: D7158457

fbshipit-source-id: e4f52cb4184ec519f30a4f11948e4633ad152c9b