Avoid parsing non-binary XML resource files
Summary:
Helper methods associated with resources assume that all XML files found in the APK will be the binary XML format as used by aapt.

This is not the case for "raw" XML files as pointed out in https://github.com/facebook/redex/issues/300 which could be a stand-alone XML file that the app parses manually at runtime. This crashes the resources parsing code here.

A "raw" XML file won't contain any resource references (thus making it uninteresting for many of the `RedexResources.cpp` helper methods), and we'll consider any class name in such a file out of scope for the renamer.

Reviewed By: justinjhendrick

Differential Revision: D7188390

fbshipit-source-id: 46b8c4645e5ccc39d4076b7938ac139c4acd2a5c