CFG remove targets and cleanup
Summary:
The edges hold all the information we need about the branch instructions and their targets. So, we can safely remove `MFLOW_TARGET` entries.

Now, while editing the CFG, we only have to maintain the edges of the CFG itself and not the targets as well.

When linearizing, the targets will be added back in.

Reviewed By: int3, arnaudvenet

Differential Revision: D7076597

fbshipit-source-id: b12a7e97895018acfabdd7c4a05d25e9101891bb