Make Redex use bionic's linux/elf.h instead of ART's
Summary: ART's `elf.h` header pulls in a ton of bionic userspace headers that are incompatible with the host's glibc toolchain. Rather than fixing up ART's and Bionic's headers to be compatible with not only the NDK toolchain but also host glibc toolchains, it's easier to point Redex code at Bionic's kernel-space ELF header which doesn't pull in any conflicts.

Reviewed By: newobj

Differential Revision: D7296956

fbshipit-source-id: e016783888cbe082d182843fa28e86dcebc3987d