remove nonsensical libresource method
Summary:
This method is causing compilation issues and it doesn't make any sense to me. It's a copy assignment operator that claims to be constant (as far as I can tell). Let's delete it because there's a non-const version right below.

Should fix these github issues:
 * https://github.com/facebook/redex/issues/307
 * https://github.com/facebook/redex/issues/298
 * https://github.com/facebook/redex/issues/286

And this gcc-7 bug report about failing to build AOSP
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=853312

Reviewed By: int3

Differential Revision: D7337841

fbshipit-source-id: 1a7d77226bc3db5c3108964ecefceef62ab86e5a