constprop: Factor into sub-analyses
Summary:
Constant propagation is useful as a "service" for several other optimizations.
The specific optimizations may want to specialize the handling of certain
instructions while leveraging the rest of the constprop logic. For example,
constprop by default treats a method's return value as unknown. However, a
specific optimization may want to do a deeper analysis of certain method calls.
With this refactoring, it can specialize the constprop logic to do just that.

Reviewed By: justinjhendrick, arnaudvenet

Differential Revision: D7391235

fbshipit-source-id: b9581286c74f39606716686eebe2bc637ceb267c