Remove problematic reflection use in instrumentation tests
Summary:
The changes I'm attempting in D7410033 make a more firm request to keep methods that are accessed by reflection.

As a result, instrumentation test behavior that is meant to inspect the output of a transform is now actually preventing the transform from taking place (lol).

Get around this by keeping behavioral verification tests in Java, and move introspection over to C++ as to not interfere.

If this smells fishy, definitely push back. Though I'm inclined to leave the logic in D7410033 as-is, since it's intuitive that reflection use should prevent some occurrences of access marking (i.e. it could be a bug if we took a method accessed via reflection and made it private).

Reviewed By: int3

Differential Revision: D7425665

fbshipit-source-id: cd9b8ace8053806004d1a01a8bb8008b8138ed23