properly encode zero branch offsets
Summary:
The Dalvik spec disallows 0 offset branch instructions. I noticed verify errors while working on a different optimization. The correct way to create a infinite loop is to place a NO-OP before the branch and go to that.

Also fixed the detection of no-op branches. The offset is measured in 16 bit code units, not `MethodItemEntry`s

Reviewed By: arnaudvenet

Differential Revision: D7432766

fbshipit-source-id: baaf9b1a9e701ae8d16d9404de29fb1b6d7b9bf6