Fix to the obfuscator to avoid renaming on present but deleted fields
Summary:
Field renaming was not taking into account collisions with existing
but deleted fields. Method renaming was already taking that into account.
I changed fields to perform that check too.
I am not clear if Wei fix would have worked too, I think so but this may be
more inline with what we do elsewhere in the obfuscator.
We don't have this yet for static fields I think which are more complicated in
the code. That may be a bit of a time bomb left behind but given I think we
should rewrite field and dmethods renaming I am leaving things as they are for
now.

Reviewed By: thezhangwei

Differential Revision: D7602211

fbshipit-source-id: 11305855308d79179289019ccbf872dd4405295f