Add parsing infrastructure to enable quickening
Summary:
This diff adds basic parsing infrastructure in oatmeal to enable
the ability to quicken opcodes.

In addition to that dexes are not streamed as a blind byte steam anymore but only as
processed by the parser who now has the ability to modify the opcodes as it processes them.

Reviewed By: dariorussi

Differential Revision: D7050351

fbshipit-source-id: 990403915371e8c25a7f51c2b7efd9e9dae387a6