D7050351: Add parsing infrastructure to enable quickening
Summary:
Original commit changeset: 6056e7576837

This diff adds basic parsing infrastructure in oatmeal to enable
the ability to quicken opcodes.
In addition to that dexes are not streamed as a blind byte steam anymore but only as
processed by the parser who now has the ability to modify the opcodes as it processes them.

Reviewed By: kgsharma

Differential Revision: D7665188

fbshipit-source-id: 90a880013bab7bf3eb9134016beb8f79069c62e1