Fix 8.1 support
Summary:
oat files for 8.1 make the information of OatDexFile a floating section within the oat
and introduce a new field in the header to point to that section. Our previous implementation was
not accounting for the 2 extra fields introduced in the OatDexFile section and was introducing
a new struct to represent that section (OatDexFileRecord) even though we already represent
that information in DexFileListing.

This diff refactors the internal model for 8.1 introducing the new fields in DexFileListing for v131 (8.1)

Reviewed By: dariorussi

Differential Revision: D7464126

fbshipit-source-id: 4b1764003e75ba6e201ab72b4d5ae0fbebb614b1