Create DisjointUnionAbstractDomain and use it for constprop
Summary:
This makes it easier to extend constant propagation to handle a variety of
different types of values. Now we don't have to add a new Environment for each
new type -- the multitude of Environments were difficult to keep in sync and
required a bunch of boilerplate code to manipulate.

Reviewed By: arnaudvenet

Differential Revision: D7690511

fbshipit-source-id: 7341a883389695ce110bfb9d238823fb6145535f