Factor out generic AbstractDomain tests
Summary:
This should make it easier to test future AbstractDomain implementations and avoid basic errors like the one in the initial implementation of D7690511.

While at it, I also defined `operator==` on AbstractDomain. This made it easier to write the generic test scaffold.

I've only done the factoring for a subset of the existing AbstractDomain tests. We can revisit the other tests in the future.

Reviewed By: arnaudvenet

Differential Revision: D7738948

fbshipit-source-id: d41200a0bef2877a6a6a803038ff91b7b8d99760