create StringBuilder for fast string concatenation
Summary:
Deobfuscate Dex Elements from ~4s to ~0.7s

These classes are necessary because `std::ostringstream` is slow to construct
and repeated calls to `std::string::operator+=` without a prior reserve
requires expensive re-allocs.

StaticStringBuilder is for a compile time known number of strings
DynamicStringBuilder is for an arbitrary number of strings
(with a runtime hint)

In practice, these seem to be fastest on a small number of strings. Too many,
and the memory cost of storing them degrades performance.

I haven't changed all the show functions, only the hot ones that appeared in profiling data.

Reviewed By: int3, minjang

Differential Revision: D7752623

fbshipit-source-id: b9a723027b7d6d338f55e0e801f118ff8ed17799