Define show() in terms of operator<< when it is available
Summary:
There are a number of classes in Redex that define their string output in terms
of the << operator. This is more convenient in some cases, but it makes it more
cumbersome to get an std::string representation of the output. Thus far I've
defined a few `str()` methods to ease this, but it was unnecessarily
repetitive. This diff handles things more generically.

Reviewed By: arnaudvenet

Differential Revision: D7809813

fbshipit-source-id: 5c41ee94dfb16b217bce31dd82514ffccafde9dd