Declare ReducedProductAbstractDomain::operator<< on the derived class
Summary:
For some reason, gtest won't print our values with the `<<` operator if it is
defined on a superclass. So I've changed the ReducedProductAbstractDomain to
define the operator directly on the derived class.

This required moving tuple_print_impl into the ReducedProductAbstractDomain
class itself, but I think that's good anyway since they were globally visible
(they were in an anon namespace but in a header file).

Reviewed By: arnaudvenet

Differential Revision: D7824942

fbshipit-source-id: 2647e4c6236f042642c2a936bc0a23a10cccbae0