Support analyzing more complicated virtual scope with a single implementation and generate single target dispatch
Summary:
This is revised version of D7770863.
We extend the change to cover the case that the interface scope only resolves to one dispatch target.
This is why we were seeing a crash on `BaseTreeModel.getGraphQLObjectType()` earlier. We observe a number of `InterfaceScope`s that only overlap with one `VirtualScope` which only has one dispatch target.

Reviewed By: emma0303

Differential Revision: D7834835

fbshipit-source-id: b238e55f862b7187a8318864aedd7988b64c1f10