Make live range renumbering into a service + make it width-aware
Summary:
The live-range-renumbering functionality of regalloc seems worth factoring out.
Ideally, we would run it as part of IRCode::balloon, but I'm not sure if it'll
break passes that rely on specific bytecode patterns (that was what happened
the last time I tried it.) However it seems worthwhile to allow certain passes
to run it, especially if they are running late in the passes list and need not
worry as much about affecting downstream passes.

Originally, live range renumbering picked the new register numbers without
regard for wide values. I.e. it would allocate v1 and v2 to separate live
ranges even if v1 denoted a wide value. This would break the IRTypeChecker (and
various passes) which assume that a wide value occupies two consecutive
registers. I have therefore added a flag to toggle this behavior -- eventually
we should make the typechecker and all our passes treat registers as abstract
IDs with no widths, but this makes the transition less painful.

Reviewed By: arnaudvenet

Differential Revision: D7962430

fbshipit-source-id: 8237ad90ce55bd7fb8872206570bb288da1af4a4