Cosmetic changes to constprop APIs
Summary:
1. Replace ConstantEnvironment::{get_primitive, get_singleton, ...} with a generic
template member function. This makes it easier to extend ConstantEnvironment
without sacrificing much readability.

2. I also defined `get_constant` directly on SignedConstantDomain, since it's one
of the most common operations that we do with that domain.

3. Renamed ConstantFieldEnvironment to StaticFieldEnvironment, which better reflects
its use. Eventually I would like to put all the stuff in ConstantEnvironment.h under
the constant_propagation namespace, so the `Constant` prefix will be unnecessary.
I also plan to use the StaticFieldEnvironment to model all static fields, not just the ones
in the current <clinit>.

Reviewed By: arnaudvenet

Differential Revision: D7963544

fbshipit-source-id: 5a55125b87ef0898923575337c8400a05c819380