assert fail when try regions are too large
Summary:
The Dalvik spec for a [try_item](https://source.android.com/devices/tech/dalvik/dex-format#try_item) uses a 16 bits for the instruction count. In the
IR, we had a try region with more than 2^16 code units in it. We downcast this
`uint32_t` to a `uint16_t`, thus leaving many instructions out of the try
region that should have been inside it!

Assert false so we can't do that in the future. I expect this assertion would
trigger if I landed this diff alone, so I plan to stack a diff on top of this
that splits the large regions into smaller ones.

See T28922924 for more details

Reviewed By: int3

Differential Revision: D7885040

fbshipit-source-id: ca296ea210de61426b4b88d1607c26e651df5451