Use classes that have mix_mode flag set
Summary:
As title. Leaving the old mechanism (with the provided list) in too,
and choosing the ones with the flag set if the file was not provided. There
still need to be other changes in (intersection between this classes and cold
start classes etc). This is just the starting point of using the analyzed
classes.

Other changes:
* TypeErasure Interdex Plugin excludes classes used for mix mode
* Added a way to enable classes from moving from the coldstart set into the mixed mode dex

Reviewed By: dariorussi

Differential Revision: D8131162

fbshipit-source-id: 17c830617b48fed6408d242912faf1b054589c60