More conservative MAX_INSTRUCTION_SIZE
Summary:
Please see T29802958.
In messenger, inliner is producing a large method that doesn't verify on 5 devices.
I'm tuning down the `MAX_INSTRUCTION_SIZE` limit as a fix.
Note that this is a config dependent soft limit and is not respected on fb4a. So I don't expect this to impact fb4a code size.
My local Redex-stats only shows about 20KB size regression for messenger.

Reviewed By: int3

Differential Revision: D8245587

fbshipit-source-id: 073f07a544974441b0da1b666bb9eccee76311b2