Finish set up for mixed mode
Summary:
This is generating a file for mixed mode that the state machine in dextricks will read to decide what DEX to compile.
We save the canary class name and use that as a DEX identifier.
This also enables mixed mode in a rather aggressive way (remove from cold start list) however is redex unstable and we can change it any time.
The code is a hack on interdex as usual

Reviewed By: emma0303

Differential Revision: D8316293

fbshipit-source-id: adfe8da905d07c82566322555261206bc161878e