Add flag for art only builds
Summary:
Added flag for the case when the build is art only.

NOTE: The way of determining if the target is art only or not, is by checking if the build rule contains `_art_` in the name. Same as how we re doing it for verifynone mode and when deciding which betamap to use. Open to other ways if there is a better way (didn't find one).

Changes:
* pass --is-art-build to `fb-redex` if the build rule has `_art_` in the name
* pass --is-art-build to `redex-all`
* pass `art_build` to pass manager
* For passes that should only be applied to ics build, check if it is an art build or not

Reviewed By: justinjhendrick

Differential Revision: D8267489

fbshipit-source-id: 2f772e7e325f82ab9e7cdc33073ce7a5279c735c