DedupBlocks on the Editable CFG
Summary:
Migrate DedupBlocksPass to use the new, editable version of the ControlFlowGraph.
It actually simplifies some of the logic too, so that's nice :)

Well, because Statsbot is being a jerk, I ran a stats run on my own machine.

It looks like a 150k win!
* Bytes. Before: 93,968,224 || After: 93,818,924 || Difference: -149,300
* Instructions. Before: 10,744,368 || After: 10,703,105 || Difference: -41,263

{F128686616}

I think the win comes from relaxing the constraint on fallthrough blocks.

Reviewed By: int3

Differential Revision: D7361560

fbshipit-source-id: 643f27fad7cdb59390e568154c1dce0d2a0effdc