Handle interface invoke in ImmutableGettersPass
Summary:
This diff analyzes GraphQL interfaces and their implementations to find interface getter invocations that can be optimized.

The pretext here is that nothing currently stops a developer from providing a hand written implementation of an autogenerated GraphQL interface (much to our dismay).

As a result, do this rather brute force analysis to collect all implementations of GraphQL interfaces transitively, and build up a set of known interface method refs which all implementations are marked as immutable.

Reviewed By: thezhangwei

Differential Revision: D8268182

fbshipit-source-id: c380cf0281ebe7246348a11fa9ab6493335587b9