dcev2: Local pointers analysis (escape analysis) [1/4]
Summary:
I am working on a version of DCE that will identify dead writes to
objects that never escape a given method. This is the first step -- identifying
the non-escaping objects.

Initially, I tried to use the constprop framework (with its escape analyzer)
for this purpose, but I realized that the current implementation was actually
unsound -- since it's only able to model one pointer per register, doing a join
of two different AbstractHeapPointers would set the register's binding to Top.
Then, when that register gets used in an escaping context (such as by being
passed to another method via an invoke instruction), we aren't correctly
escaping the values bound to those pointers. This analysis handles this
properly by tracking *all* possible register values.

Reviewed By: justinjhendrick, arnaudvenet

Differential Revision: D8366559

fbshipit-source-id: 4e61c990463dd64da49eb07c4a447f370cd70916