assumenosideeffects on external classes
Summary:
We weren't parsing any proguard rules on external classes. This diff fixes that

I copied most of this (with permission) from Brian's D8462533 which he never ended up landing.

To choose which additional methods to assumenosideeffects on, I printed out which external methods LocalDce was considering for purity and sorted by number of occurrences. The methods I added were called at least 1000 times and were unlikely to cause problems if elided.

O, and it's a 75k win!

Reviewed By: wsanville

Differential Revision: D8500360

fbshipit-source-id: 538995d85c0cae3f512510a9f1ad36163cfec0c2