Rewrite interface dispatch cleanup using the CFG api
Summary:
Instead of scanning through the instruction stream and try matching instruction patterns, we can manipulate the CFG and clean up repeating blocks at a higher level.
This is enabled by the recent change on the CFG API.

We essentially do the following steps. First we identify invoke blocks that calls a repeating target and remove them. Then we clean up the CFG by removing blocks that become unreachable or pass-through block that can be short circuited.

Reviewed By: emma0303, justinjhendrick

Differential Revision: D8528060

fbshipit-source-id: 46beb5f765910011cee4807921327876d222297f