Fix cfg::Edge::operator==
Summary:
We weren't considering the case_key while checking the equality of edges.

This manifest as a bug where DedupBlocks merged two blocks that had the same
outgoing edges but for different cases of a switch statement.

I'll copy to staging and unstable in a separate diff.
I'll re-enable for instagram in a separate diff.

Reviewed By: emma0303

Differential Revision: D8586709

fbshipit-source-id: 8dd60b2f8d7e17457c6c505543de1726ceba7982