FinalInlineV2
Summary:
This re-implements FinalInline using the constprop machinery. The previous
implementation relied on bytecode pattern-matching to determine constant
fields, which caused it to miss various cases as well as to wrongly identify
certain fields as constant.

In theory, all this could be done as part of IPCP, but there are a few
advantages to doing it as a separate pass:

* It's much faster to run than full IPCP
* We may want to run this pass early in the pass list to make other passes more
  effective, then follow up with full IPCP later on

Reviewed By: justinjhendrick, arnaudvenet

Differential Revision: D8294112

fbshipit-source-id: 2dfd5872fba9b48632be40a75da659ce8bd6aadc