LocalDce on the new CFG
Summary:
Have LocalDce operate on an editable CFG instead of editing IRCode and rebuilding.

No significant change in APK size but [the statsbot run](https://our.intern.facebook.com/intern/redex/stats/?handle=GOHxFwK34p3HnRUkAAAAAADrjAZIbqsmAAAz) shows decreases in LocalDce's stats. This is because the CFG removes unreachable blocks when it is constructed (at the beginning of LocalDce), so some block removals are no longer being reported in LocalDce's stats. This isn't the most accurate for now, but in the long run (when the CFG is created once and destroyed once), this will be the correct way to measure

Reviewed By: int3

Differential Revision: D8309621

fbshipit-source-id: 53e335c28c7e8b55f976c37b6fc2ebde47dbce66