Add Disjoint interface analysis
Summary:
This analysis pass finds interfaces that are "disjointly" implemented.

More precisely, we want to find an interface I and a set of classes C such that:
  1. each class in C implements and only implements I
  2. Classes in C share the same super class (or impilicitly java.lang.Object)

If the cardinality of C is large, we could potentially apply type erasure on C.

One example of such set of classes is anonymous PARCELABLE.CREATOR classes.

Reviewed By: thezhangwei, emma0303

Differential Revision: D8667659

fbshipit-source-id: f1aa6b0fb2209c3d1170f3ffa7a4ca4a5c3d8bb9