Fix OSS build issues around jemalloc linking
Summary: It was broken by D8701633, which tried to enable us to call `mallctl` without making jemalloc a build dependency. I tried to handle that by using weak linking, but that seems to be implemented slightly differently across Linux and MacOS, and I couldn't find a way to make it work well with both. I'm switching to using dynamic symbol lookup here which hopefully works in a more portable way.

Reviewed By: minjang

Differential Revision: D8711857

fbshipit-source-id: b6ff157d179960e0ef0ac35f0a42603ee0ac76b1