DexEndMarker0.class -> LDexEndMarker0;
Summary:
As title.

For Instagram, since we are touching the first dex, we want to make sure that we also add any primary dex class that was missing from the interdex order to it. We use DexEndMarker0 to identify the delimitation between first and second dex. This fix gets the correct delimiter. Before that, since we weren't getting the correct end of the first dex, we might have ended up putting primary dex classes in one of the next dexes. NOTE: This wasn't used yet.

Reviewed By: minjang

Differential Revision: D8716463

fbshipit-source-id: 6f2c2d2664e353078fc227adf1e79ba5c442acb5