Allow for custom equality check for IRInstruction
Summary:
This diff is meant to allow for reusing all of the code in `structural_equals()` for comparing branching, try/catch etc, but allow for a custom equality function for two instructions.

This will be used in a later diff to compare anonymous inner classes for equality (which uses a special check to abstract away fields of the same type in different classes).

This might not be the most idiomatic way to do this, but wanted to make this optional and leave all existing call sites the same.

Reviewed By: int3, emma0303

Differential Revision: D8776286

fbshipit-source-id: e8e0f788cb6c72fdaafabe3f6035ad1c1cd54a4d