Static method relocation
Summary:
We figured out there are 400+ classes that only contain static methods which are can_rename() and can_delete() and referenced only once. Relocating these methods could make RemoveEmptyClasses be able to remove more classes .

Implementation
Instead of relocating a single static method that are only referenced once, I try to relocate a group of static methods at the same time, if the group of methods only have one entry point from a non-static method. Here is the overview of the algorithm.

  1. Generate candidate class set that only contain static methods
  2. Build call graph for all these static methods in candidate class set
  3. For each other classes, find out the invocations to these static methods transitively, color these static methods into the class's color. In another word, I figure out the clusters of static methods that may be able to relocated to this class.
  4. Relocate based on the coloring results.
     a. If a static method only has 1 color, relocate it to the class of that color
     b. If a static method has no color, relocate it if it has only one caller, log it if it has no caller.
     c. If a static method has multiple colors, keep it untouched.

Reviewed By: emma0303

Differential Revision: D8853686

fbshipit-source-id: e315bda33077460fda709e0664ece0e11fa479a9