Make escape analysis interprocedural
Summary:
Without interprocedural analysis, we have to assume that every method call
causes all its arguments to escape, which gives our analysis pretty poor
precision. This diff addresses that problem by generating summaries for each
method it analyzes. These summaries indicate which of the input arguments end
up escaping.

The summary also records if the return value of a method is one of the input
parameters. This is particularly useful for methods that return `this`, such as
`StringBuilder.append`.

Reviewed By: justinjhendrick

Differential Revision: D9073010

fbshipit-source-id: 6e0088c6d3dc1a5fdee37c44e42b6ad7dfa69718