add jemalloc and profiling util srcs to cmake build (#338)
Summary:
This commit updates the cmake build script to include
`util/CommandProfiling` and `util/JemallocUtil` as a fix for
missing symbols when building with cmake.

These already appear in Makefile.am so I expect the cmake src list has
just fallen out of date.

I used `.*` to also include the headers in the source list for consistency
with `utils/Sha1.*` which was already present. All the other dirs
in that section just include `*.h` and `*.cpp` but I'm assuming you want to
avoid MallocDebug for normal builds.
Pull Request resolved: https://github.com/facebook/redex/pull/338

Reviewed By: minjang

Differential Revision: D9183133

Pulled By: thezhangwei

fbshipit-source-id: 1f066316520edbf161d4cd37ae358522cdd6bda3