dcev2: Parallelize side-effect analysis [1/6]
Summary:
I've refactored the code in a similar style to the parallelized escape analysis
in {D9073010}. The per-method side effect analyzer now takes an
InvokeToSummaryMap instead of trying to resolve methods on its own, allowing us
to switch between different call graph analyses.

This diff also puts the side-effect analysis code under a namespace. This
unfortunately accounts for many churn lines in the diff that were hard to split
out from the rest of the changes.

Reviewed By: justinjhendrick

Differential Revision: D9176363

fbshipit-source-id: 732a7465286f8dd426a684ed023bfc6926529a99