dcev2: Generate escape summaries for external methods [2/6]
Summary:
This diff extends the summary-generator tool to emit both side-effect and
escape analysis summaries. It also extends the predefined StringBuilder
analysis results. StringBuilder method calls are one of the larger sources of
dead code and having a precise model of their effects allows us to remove more
of them.

Reviewed By: justinjhendrick

Differential Revision: D9176382

fbshipit-source-id: 10b27303e03885496118364b4b120cbdb6caa0c5