Change to Vector based approach for basic block access evaluation
Summary:
Here, we change our approach from instrumenting individual blocks to having a per method metric to measure the basic block access. This makes minimal changes to the original CFG.

The current version works only for methods with 32 basic blocks or less. This covers about 92% of the methods in fb4a.
I will work on extending this approach to more methods with more than 32 basic blocks.

Reviewed By: minjang

Differential Revision: D9074512

fbshipit-source-id: 1984419da7ed1e762acaa651fbad0596cb218644