Set plugin generated classes as emitted
Summary:
As title.

While working on the next diff (to come), realized we don't keep track (as in
add to emitted) the classes that were returned by the plugins. This worked
well, since **most** of the classes, should be generated and/ or no error
appears even if those are not added since they will be skipped, if the class is
found again.

To work around this issue, changes:
* keep track of those emitted classes
* in TypeErasureInterdexPlugin we might end up with a case where
** class B is selected to be merged (but not skipped yet)
** we see that class B cannot be merged -> we add it in the
m_mergeables_still_emitted and return it together with additional classes
** when we do find class B in the scope, we will simply ignore it (since it was
emitted, so there is nothing left to skip)
** update the cleanup assert to check that all classes that appeared in merged
are also accounted for in m_mergeables_still_emitted. (we could have removed
them from merged too, but this seems a bit cleaner to me)

Reviewed By: thezhangwei

Differential Revision: D9154149

fbshipit-source-id: d860103cd7687cda60bbc7aedf124a6852b6033d