Refactor
Summary:
As title. I tried to make the diff smaller, but once I seperated the dexes
structure, I kept touching most of the code, and in the end here we are. With
no change in the logic of the algorithm, I went with this, instead of trying to
separate more.

Changes:
* moved dex structure in it's own class(es). Everything related to what fits where is handeled by DexStructure + DexesStructure
** NOTE: We only keep track of the latest dex, while constructing it, but can easily be extended.

* InterDex takes care of the logic around primary dex / interdex order / what gets mixedmode etc

Pros:
* no more global variables :D
* no more huge methods, hard to debug and update :D
* less redundant code
* easier to maintain and update later on

Reviewed By: int3

Differential Revision: D9163452

fbshipit-source-id: 442251f989aef5c9cc54c7d9afc9d571ebd9dfad