Extend Basic Block Instrumentation to methods with maximum 80 blocks
Summary:
Methods with less than 80 basic blocks are instrumented using a maximum of 5 16-bit bit vectors.
To do this, the final analysis method is overloaded to accept variable(1,2,3,4,5) number of bit vectors.
In the final step of the algorithm, we will cover the case of more than 5 bit vectors(i.e. more than 80 basic blocks).

Reviewed By: minjang

Differential Revision: D9256390

fbshipit-source-id: 5050f4f7734417177e974358644800f8762a65af