Port check that we have mixedmode classes
Summary:
As title. Same as D9031595.

For instagram and fblite, this will make a change in the way their last
coldstart dex is terminated.

Usually, this method wasn't suppose to do anything since there aren't any
mixedmode classes and this would have been an empty dex. For fblite though,
since they don't break after the interdex order -> we end up splitting their
secondary-1 into 2 dexes. This fixes that.

NOTE: This is fixed in unstable and staging with the refactoring. I would look into porting it to stable sooner than the quarterly (if it wouldn't imply to many changes).

Reviewed By: int3

Differential Revision: D9319433

fbshipit-source-id: 284571eb48214f6145e181d58b190039cfcfecc3