Fix colliding method by appending variable additional args
Summary:
When updating type references on a method signature, the updated signature could collide with an existing one.
The current solution is to artificially append an additional parameter to the method to make them different and patch the call site by passing an additional constant argument.
This doesn't work if we have more than two methods colliding with each other.
The diff proposes a potential fix for the limitation.

Reviewed By: emma0303

Differential Revision: D9302391

fbshipit-source-id: 74e299aa72c4319130d14b5237526659c3d54b4e