delete unreferenced edges
Summary:
`cfg-inliner-test` was intermittently failing with some memory related error. I traced it down to the fact that we had an edge that was referencing a non-existent block. This made me realize that removing an edge from the graph, but not deleting it, is a bug. So I scoured `ControlFlow.cpp` for that behavior and found those instances.

In an attempt to prevent this in the future, `simplify()` now will also remove unreferenced edges.

Also, rename `remove_all_edges` to `remove_edges_between` because it's more descriptive and less misleading.

Reviewed By: arnaudvenet

Differential Revision: D9415095

fbshipit-source-id: c7a6188bd7e331348ecac92c5587c2c900824aab