Keep running tally of opcode size when inlining
Summary:
There seems to be a unexpected case where a method that has < 3 opcodes is marked as always inlinable.

However, a larger method might get inlined to the small method, leading to code duplication. Saw this happening in a few cases.

Unit test should illustrate the problem, before the patch the code ends up with `bigMethod` duplicated twice inside `testUpdateCodeSizeWhenInlining`.

Reviewed By: int3, justinjhendrick

Differential Revision: D9390218

fbshipit-source-id: 4137667f1b4c2cb6422ef0d5a77071eb5af11ea6