Change dynamic task list API of WorkQueue
Summary:
The original API required the parallelized job to capture the WorkQueue by reference in order to add tasks to it while running. This was not ideal for two reasons:

1. Instead of being able to initialize the WorkQueue in one line, you had to initialize it first with a dummy mapping function, then set the real mapping function later on. This two-step process was necessary in order for the mapping function to capture a reference to the WorkQueue.
2. `WorkQueue::add_item()` assigns a new task to a random worker, which means that it needs to acquire a lock on the queue that the worker is consuming tasks from. Ideally, a job should be adding tasks to its own queue to minimize lock contention. (In practice our WorkQueue implementation doesn't seem optimized enough yet for this difference to matter in the performance of RMU though. But it's good future-proofing)

In order to do this, I needed to pass each job a handle to its corresponding WorkerState. I decided to replace the `Data` parameter with it since `Data` was almost never used. The Data object is still accessible through the WorkerState pointer though. (I also removed the parameter entirely from the signature of `reduce_methods`.)

Reviewed By: minjang

Differential Revision: D9414547

fbshipit-source-id: 23727fc5027d3dfd0ba7f47c1721b2931db22b6d