Do not generate empty dex files
Summary:
`redex-opt` may crash when loading dex files that are generated before InterDex and after TypeErasure. It turns out there are no classes in these dex. Ignore them when writing dex files.
Thanks Emma for helping me figure out this.

And the crash is because of an uninitialized pointer P60008106.

Reviewed By: minjang

Differential Revision: D9587964

fbshipit-source-id: b762597e95d5a29a269750fff3ec190ffff50d4b