Add redex-class-method-info-map
Summary:
Having information about class/method such as sizes is useful for later analysis. This extends the current "redex-method-info-map" to "redex-**class**-method-info-map".

The format is a CSV, but some common strings and classes are interned to save the space. The only downside is that we need a bit smarter script. This will be implemented in the following script (method-trace-parser.py).

Reviewed By: thezhangwei

Differential Revision: D9595725

fbshipit-source-id: a066a79167a4a8e709fd0772638e4ec87f71f179