Lift partially invalid constant values
Summary:
Currently when one of the constant value in an annotation is invalid or cannot be parsed, we skip the entire set of ConstantValues. Doing that leads to us missing some of the opportunities to dedup.
This change enables lifting the rest of the valid constant values. Therefore, enables more methods to be deduplicated.

Reviewed By: emma0303

Differential Revision: D9525462

fbshipit-source-id: d6ccf389a60e2856682b256076a2eade77099fdd