Use explicit JSON assets list file to strip JS resources
Summary:
Redex implements a logic in [RedexResources.cpp](https://fburl.com/qo0wd53k) which parses source code JS bundle, greps it for `.png` and other resources and strips the assets that end up being not referenced anywhere. This logic is quite hard to maintain and, now that we are switching to Hermes Bytecode instead of plain JS, it appears to be quite limiting for our build configuration. With Fb4a, for example, we **have** to add both JS and Bytecode bundles in the APK and then strip the former one as a part of postprocessor step which is executed after Redex (see https://fburl.com/ocminc9h). Full context is available in T11667256.

Since Metro has a full knowledge about JS assets and now with the changes implemented by @[1697624685:Jean Lauliac] in D9658242, it can output the full assets list, lets use it in Redex in order to drive resource stripping logic.

This diff introduces a `--used-js-assets` flag to the Redex CLI interface, which allows to pass an explicit file with Android asset identifiers (for example, P60031902) which we then use in order to drive resource stripping logic.

I will follow up with Fb4a integration in a next diff in the stack.

Differential Revision: D9682240

fbshipit-source-id: e673a2f975d40a930ce46011e39fa5a5fca84ecb